Isolate runtime embedding entry adapters
- Move typed embedded call entry points into a dedicated runtime owner.
- Preserve the crate-private facade and existing family-specific evaluator paths.
